PUBDEV-8843: Update java tests
